+++ packages/components/components/elvis-dropdown/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Patch
 
-- Bumped dependency
+- Updated tokens declarations
 
 ## 2.0.3 (03.01.22)